# Services/rooting/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Graph API",
    "description": "This API can create chart from dataSets"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/dataSets/": {
      "post": {
        "tags": [
          "DataSets"
        ],
        "description": "Create a dataSet",
        "parameters": [
          {
            "name": "points",
            "in": "body",
            "description": "Point of the dataSet",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/points"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Return the id of the new dataSet"
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "500": {
            "description": "Return the error message"
          }
        }
      }
    },
    "/dataSets/{dataSetId}": {
      "put": {
        "tags": [
          "DataSets"
        ],
        "description": "Modify a specific dataSet",
        "parameters": [
          {
            "name": "dataSetId",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "param",
            "description": "The id of the dataSet to modify",
            "required": true,
            "type": "integer"
          },
          {
            "name": "information to modify",
            "in": "body",
            "description": "Element of the dataSet to modify",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/dataSet"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "MODIFY"
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "500": {
            "description": "Return the error message"
          }
        }
      }
    },
    "/dataSets/{id}/": {
      "delete": {
        "tags": [
          "DataSets"
        ],
        "description": "Suppression of a specific dataSet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "idd",
            "in": "param",
            "description": "The id of the dataSet",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "Return \"The dataSet with id : {id} has been successfully deleted\""
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "500": {
            "description": "Return the error message"
          }
        }
      }
    },
    "/dataSets/timeseries": {
      "get": {
        "tags": [
          "DataSets"
        ],
        "description": "Create a timeSerie dataSet",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Return the id of the dataSet created"
          },
          "501": {
            "description": "Return the error message"
          }
        }
      }
    },
    "/dataSets/timeseries/": {
      "post": {
        "tags": [
          "DataSets"
        ],
        "description": "Modify a timeSerie dataSet",
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "description": "The id of the dataSet to modify",
            "required": true,
            "type": "integer"
          },
          {
            "name": "timestamp",
            "in": "body",
            "description": "The timestamp of the data",
            "required": true,
            "type": "integer"
          },
          {
            "name": "value",
            "in": "body",
            "description": "The value of the data (int or boolean)",
            "required": true,
            "type": "object"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "501": {
            "description": "Return the error message"
          }
        }
      }
    },
...
    "/dataSets/{id}": {
      "get": {
        "tags": [
          "DataSets"
        ],
        "description": "Get all information about a dataSet",
        "parameters": [
          {
            "name": "id",
            "in": "param",
            "required": true,
            "description": "The id of the dataSet",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Return all information about a dataSet",
            "schema": {
              "$ref": "#/definitions/dataSetRead"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "points": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "example": 750
          },
          "y": {
            "type": "number",
            "example": 602
          }
        }
      },
      "xml": {
        "name": "points"
      }
    },
    "dataSet": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "dataSet name"
        },
        "description": {
          "type": "string",
          "example": "dataSet description"
        },
        "points": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "label": {
                "type": "string",
                "example": "May"
              },
              "value": {
                "type": "number",
                "example": 190
              }
            }
          }
        }
      },
      "xml": {
        "name": "dataSet"
      }
    },
    "dataSetRead": {
      "type": "object",
      "properties": {
        "iddataset": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "myDataSet"
        },
        "description": {
          "type": "string",
          "example": "It's my dataSet"
        },
        "creationdate": {
          "type": "string",
          "example": "2020-12-10T08:02:06.680Z"
        },
        "datajson": {
          "type": "string",
          "example": "[{\"label\":\"January\",\"value\":20},{\"label\":\"February\",\"value\":60},{\"label\":\"March\",\"value\":70},{\"label\":\"April\",\"value\":180},{\"label\":\"May\",\"value\":290}]"
        }
      },
      "xml": {
        "name": "dataSetRead"
      }
    },
...
  }
}